Xeyem/requirements.txt
beautifulsoup
free-proxy
pandas
plotly
google
django-extensions
